rename components to index
