+++ src/app/(auth)/sign-up/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useRef, useEffect, useTransition, FormEvent } from "react";
 import { useSearchParams } from "next/navigation";
 
-import Input, { type InputType } from "@/components/input/input";
+import Input, { type InputType } from "@/components/input";
 import Button, { type ButtonType } from "@/components/button";
 
 import { SignUpAction } from "@/actions/auth.actions";
